Extract JavaToKotlinConverter
